requirements.txt
PyQt5>=5.15.9
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.11.0
requests>=2.31.0
Pillow>=10.0.0
pyinstaller>=5.13.0
